pyproject.toml
[tool.poetry]
name = "upscaler"
version = "1.0"
description = "An image upscaling library with a unified interface for multiple model architectures."
authors =  ["Sean Norini <snorini@gmail.com>"]
readme = "README.md"
packages = [{ include = "upscaler" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
numpy = "<2.3.0"
spandrel-extra-arches = "^0.2.0"
opencv-python = ">=4.12.0.88,<5.0.0.0"
pytest-mock = "^3.15.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"